+++ sdk/search/azure-search-documents/src/samples/java/com/azure/search/documents/SearchOptionsAsyncExample.java
@@ -3,6 +3,7 @@
 
 package com.azure.search.documents;
 
+import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.util.Configuration;
 import com.azure.search.documents.models.FacetResult;
 import com.azure.search.documents.models.RequestOptions;
@@ -37,7 +38,7 @@ public class SearchOptionsAsyncExample {
     public static void main(String[] args) {
         SearchIndexAsyncClient searchClient = new SearchIndexClientBuilder()
             .endpoint(ENDPOINT)
-            .credential(new SearchApiKeyCredential(API_KEY))
+            .credential(new AzureKeyCredential(API_KEY))
             .indexName(INDEX_NAME)
             .buildAsyncClient();